dap_status becomes dap::core_status
